refactor(#1440): move async wrappers from core/ to services/
PR #1482 merged without this fix. Move 3 async wrapper implementations
out of core/ (kernel boundary violation):
- async_agent_registry.py → services/agents/
- async_namespace_manager.py → services/permissions/
- async_router.py → services/routing/

Update all imports in fastapi_server.py, unit tests, and e2e tests.
