+++ src/nexus/services/agents/async_agent_registry.py
