+++ tests/unit/services/test_async_router.py
@@ -6,9 +6,9 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from nexus.core.async_router import AsyncVFSRouter, _to_mount_info, _to_resolved_path
 from nexus.core.protocols.vfs_router import MountInfo, ResolvedPath, VFSRouterProtocol
 from nexus.core.router import MountConfig, PathNotMountedError, RouteResult
+from nexus.services.routing.async_router import AsyncVFSRouter, _to_mount_info, _to_resolved_path
 from tests.unit.core.protocols.test_conformance import assert_protocol_conformance
 
 # ---------------------------------------------------------------------------
